refactor: myactivities 도메인 폴더 삭제
