--- src/domain/myactivities/type.ts
@@ -1,20 +1,0 @@
-export interface Activity {
-  id: number;
-  userId: number;
-  title: string;
-  description: string;
-  category: string;
-  price: number;
-  address: string;
-  bannerImageUrl: string;
-  rating: number;
-  reviewCount: number;
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface MyActivities {
-  cursorId: number;
-  totalCount: number;
-  activities: Activity[];
-}
